LayoutTests/platform/qt-4.8/fast/css/ex-unit-with-no-x-height-expected.txt

This test ensures that the CSS 'ex' unit is treated as half the size of the CSS 'em' unit when a font has no x-height information.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL measureExBox(); should be 50px. Was 45px.
PASS successfullyParsed is true

TEST COMPLETE
